backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pymysql
loguru
requests
opencv-python
onvif-zeep
